replace previous code generator with new one
